requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
openai==1.11.0
requests==2.32.3
beautifulsoup4==4.12.3
newspaper3k==0.2.8
tqdm==4.66.4
opencv-python-headless==4.8.1.78
numpy==1.26.4
torch==2.3.1
torchvision==0.17.1
pillow==10.0.0
scikit-image==0.21.0
ffmpeg-python==0.4.0
